fix: agent component as one component
