+++ apps/comps/components/agent-profile/index.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import {Share2Icon} from "lucide-react";
 import React from "react";
 
 import Card from "@recallnet/ui2/components/card";
@@ -16,7 +15,6 @@ import {
 } from "@recallnet/ui2/components/table";
 import {
   Tabs,
-  TabsContent,
   TabsList,
   TabsTrigger,
 } from "@recallnet/ui2/components/tabs";
@@ -24,9 +22,15 @@ import {cn} from "@recallnet/ui2/lib/utils";
 
 import {Hexagon} from "@/components/hexagon";
 import MirrorImage from "@/components/mirror-image";
-import {Agent, AgentWithOwnerResponse, Competition, CompetitionStatus, CrossChainTradingType} from "@/types";
+import {Agent, AgentWithOwnerResponse, Competition, CompetitionStatus} from "@/types";
 import {BreadcrumbNav} from "../breadcrumb-nav";
 import {useAgentCompetitions} from "@/hooks/useAgentCompetitions";
+import {useUpdateAgent, useUserAgents} from "@/hooks";
+import {ShareAgent} from "./share-agent";
+import {AgentImage} from "./agent-image";
+import {EditAgentField} from "./edit-field";
+import AgentInfo from "./agent-info";
+import CompetitionTable from "./comps-table";
 
 export default function AgentProfile({
   id,
@@ -46,6 +50,9 @@ export default function AgentProfile({
 }) {
   const skills = agent?.skills || [];
   const trophies = (agent?.trophies || []) as string[];
+  const {data: userAgents} = useUserAgents();
+  const isUserAgent = userAgents?.agents.some((a) => a.id === id) || false;
+  const updateAgent = useUpdateAgent();
 
   const sortString = React.useMemo(() => {
     return Object.entries(sortState).reduce((acc, [key, sort]) => {
@@ -54,33 +61,21 @@ export default function AgentProfile({
     }, "");
   }, [sortState]);
 
-  const comp = {
-    id: 'id',
-    name: "name",
-    description: "description",
-    externalUrl: "externalUrl",
-    imageUrl: "imageUrl",
-    type: "type",
-    status: CompetitionStatus.Active,
-    crossChainTradingType: CrossChainTradingType.Allow,
-    startDate: null,
-    endDate: null,
-    createdAt: "createdAt",
-    updatedAt: "updatedAt",
-    stats: {
-      totalTrades: 0,
-      totalAgents: 0,
-      totalVolume: 0,
-      uniqueTokens: 0,
-    },
-    votingEnabled: true,
-    userVotingInfo: {
-      canVote: true,
-      info: {
-        hasVoted: true,
+  const handleSaveChange =
+    (field: "imageUrl" | "description" | "name") => async (value: unknown) => {
+      if (!agent) return;
+
+      try {
+        await updateAgent.mutateAsync({
+          agentId: agent.id,
+          params: {
+            [field]: value,
+          },
+        });
+      } catch (error) {
+        console.error("Failed to update agent:", error);
       }
-    }
-  }
+    };
 
   const {data: compsData} =
     useAgentCompetitions(id, {sort: sortString, status});
@@ -104,30 +99,54 @@ export default function AgentProfile({
           cropSize={45}
         >
           <div className="flex w-full justify-end">
-            <Share2Icon className="text-gray-600" size={30} />
+            <ShareAgent agentId={agent.id} />
           </div>
-          <MirrorImage
-            image={agent.imageUrl || "/agent-placeholder.png"}
-            width={160}
-            height={160}
-          />
+          {isUserAgent ?
+            <AgentImage
+              agentImage={agent?.imageUrl || "/agent-placeholder.png"}
+              onSave={handleSaveChange("imageUrl")}
+            />
+            :
+            <MirrorImage
+              image={agent.imageUrl || "/agent-placeholder.png"}
+              width={160}
+              height={160}
+            />
+          }
           <span className="w-50 mt-20 text-center text-lg text-gray-400">
             Calm accumulation of elite assets.
           </span>
         </Card>
         <div className="flex-2 xs:col-span-2 xs:col-start-2 xs:row-start-1 xs:mt-0 xs:h-[65vh] col-span-3 row-start-2 mt-5 flex shrink flex-col border lg:col-span-1 lg:col-start-2">
-          <div className="grow border-b p-8">
-            <h1 className="truncate text-4xl font-bold text-white">
-              {agent.name}
-            </h1>
-            <div className="mt-5 flex w-full gap-3">
-              <span className="text-xl font-semibold text-gray-400">
-                Developed by
-              </span>
-              <span className="truncate text-xl font-semibold text-gray-400 text-white">
-                {owner?.name}
-              </span>
-            </div>
+          <div className="grow border-b p-8 relative">
+            {
+              isUserAgent ?
+                <EditAgentField
+                  title="Agent Name"
+                  value={agent.name || ""}
+                  onSave={handleSaveChange("name")}
+                >
+                  <h1 className="max-w-[90%] truncate text-4xl font-bold text-white">
+                    {agent.name}
+                  </h1>
+                </EditAgentField>
+
+                :
+                <h1 className="truncate text-4xl font-bold text-white">
+                  {agent.name}
+                </h1>
+            }
+            {
+              !isUserAgent &&
+              <div className="mt-5 flex w-full gap-3">
+                <span className="text-xl font-semibold text-gray-400">
+                  Developed by
+                </span>
+                <span className="truncate text-xl font-semibold text-gray-400 text-white">
+                  {owner?.name}
+                </span>
+              </div>
+            }
             <div className="mt-8 flex w-full justify-start gap-3">
               {trophies.length > 0 ? (
                 trophies.map((_: unknown, i: number) => (
@@ -142,6 +161,10 @@ export default function AgentProfile({
                 </span>
               )}
             </div>
+            {
+              isUserAgent &&
+              <div className="absolute bottom-5 w-full max-w-[500px]"><AgentInfo agent={agent} /></div>
+            }
           </div>
           <div className="flex flex-col items-start gap-2 border-b px-6 py-12 text-sm">
             <span className="w-full text-left font-semibold uppercase text-gray-400">
@@ -174,9 +197,21 @@ export default function AgentProfile({
         </div>
         <div className="xs:grid col-span-3 row-start-2 mt-8 hidden grid-rows-2 border-b border-l border-r border-t text-sm lg:col-start-3 lg:row-start-1 lg:mt-0 lg:h-[65vh] lg:grid-rows-3 lg:border-l-0">
           <div className="flex flex-col items-start gap-2 border-b p-6 lg:row-span-2">
-            <span className="font-semibold uppercase text-gray-400">
-              agent description
-            </span>
+            {isUserAgent ?
+              <EditAgentField
+                title="Agent Profile"
+                value={agent.description || ""}
+                onSave={handleSaveChange("description")}
+              >
+                <span className="font-semibold uppercase text-gray-400">
+                  agent description
+                </span>
+              </EditAgentField>
+              :
+              <span className="font-semibold uppercase text-gray-400">
+                agent description
+              </span>
+            }
             <span className="text-gray-400">
               {agent.description || "No profile created yet"}
             </span>
@@ -258,7 +293,7 @@ export default function AgentProfile({
           <CompetitionTable
             handleSortChange={handleSortChange}
             sortState={sortState}
-            isLoading={false}
+            canClaim={isUserAgent}
             competitions={competitions || []}
           />
         </Tabs>
@@ -267,155 +302,3 @@ export default function AgentProfile({
   );
 }
 
-function CompetitionTable({
-  competitions,
-  handleSortChange,
-  sortState,
-  isLoading,
-}: {
-  competitions: Competition[] | undefined;
-  handleSortChange: (field: string) => void;
-  sortState: Record<string, SortState>;
-  isLoading: boolean;
-}) {
-  console.log({isLoading})
-  return (
-    <div className="overflow-hidden rounded border">
-      <Table>
-        <TableHeader className="text-muted-foreground bg-gray-900 text-xs uppercase">
-          <TableRow className="grid w-full grid-cols-7">
-            <SortableTableHeader
-              onToggleSort={() => handleSortChange("name")}
-              sortState={sortState["name"]}
-            >
-              Competition
-            </SortableTableHeader>
-            <SortableTableHeader
-              onToggleSort={() => handleSortChange("skills")}
-              sortState={sortState["skills"]}
-            >
-              Skills
-            </SortableTableHeader>
-            <SortableTableHeader
-              onToggleSort={() => handleSortChange("portfolio")}
-              sortState={sortState["portfolio"]}
-            >
-              Portfolio
-            </SortableTableHeader>
-            <SortableTableHeader
-              onToggleSort={() => handleSortChange("pnl")}
-              sortState={sortState["pnl"]}
-              className="w-30 flex justify-end"
-            >
-              P&L
-            </SortableTableHeader>
-            <SortableTableHeader
-              onToggleSort={() => handleSortChange("trades")}
-              sortState={sortState["trades"]}
-            >
-              Trades
-            </SortableTableHeader>
-            <SortableTableHeader
-              onToggleSort={() => handleSortChange("placement")}
-              sortState={sortState["placement"]}
-            >
-              Placement
-            </SortableTableHeader>
-            <TableHead className="text-left">Trophies</TableHead>
-          </TableRow>
-        </TableHeader>
-
-        <TableBody>
-          {!isLoading && competitions && competitions.length > 0 ? (
-            competitions.slice(0, 10).map((comp, i) => {
-              const compStatus =
-                comp.status === CompetitionStatus.Active
-                  ? {
-                    text: "On-going",
-                    style: "border-green-500 text-green-500",
-                  }
-                  : comp.status === CompetitionStatus.Pending
-                    ? {
-                      text: "Upcoming",
-                      style: "border-blue-500 text-blue-500",
-                    }
-                    : {
-                      text: "Complete",
-                      style: "border-gray-500 text-gray-500",
-                    };
-
-              return (
-                <TableRow key={i} className="grid grid-cols-7">
-                  <TableCell className="flex flex-col justify-center">
-                    <span className="truncate text-sm font-semibold text-gray-400">
-                      {comp.name}
-                    </span>
-                    <span
-                      className={cn(
-                        "mt-1 w-fit rounded border px-2 py-0.5 text-xs font-medium",
-                        compStatus.style,
-                      )}
-                    >
-                      {compStatus.text}
-                    </span>
-                  </TableCell>
-                  <TableCell className="flex flex-wrap items-center gap-2">
-                    {/* Future skills mapping */}
-                  </TableCell>
-                  <TableCell className="text-md flex items-center font-medium text-gray-400">
-                    $0<span className="ml-2 text-xs">USDC</span>
-                  </TableCell>
-                  <TableCell className="w-30 flex items-center justify-center font-medium">
-                    <span className="flex flex-col text-gray-400">0$</span>
-                  </TableCell>
-                  <TableCell className="w-30 text-md fond-semibold flex items-center text-center text-gray-400">
-                    0
-                  </TableCell>
-                  <TableCell className="w-30 flex items-center text-center text-gray-400">
-                    0/0
-                  </TableCell>
-                  <TableCell className="align-center h-25 flex items-center gap-2">
-                    <Hexagon className="h-8 w-8 bg-blue-500" />
-                    <Hexagon className="h-8 w-8 bg-green-500" />
-                    <Hexagon className="h-8 w-8 bg-yellow-500" />
-                  </TableCell>
-                </TableRow>
-              );
-            })
-          ) : (
-            <TableRow>
-              <TableCell colSpan={7} className="p-5 text-center">
-                {
-                  isLoading ?
-                    <div>LOADING</div>
-                    :
-                    <div className="flex flex-col">
-                      <span className="font-bold text-gray-400">
-                        This agent hasn’t joined any competitions yet
-                      </span>
-                      <span className="text-gray-600">
-                        Participated competitions will appear here once the agent
-                        enters one.
-                      </span>
-                    </div>
-                }
-              </TableCell>
-            </TableRow>
-          )}
-        </TableBody>
-      </Table>
-
-      {competitions && competitions.length > 10 && (
-        <div className="max-h-[600px] overflow-y-auto">
-          <Table>
-            <TableBody>
-              {competitions.slice(10).map((comp, i) => (
-                <TableRow key={i}>{/* Same structure */}</TableRow>
-              ))}
-            </TableBody>
-          </Table>
-        </div>
-      )}
-    </div>
-  );
-}
